Modified wasm div function

+++ lang/javascript/learningmachinelearning/assembly/module.ts
@@ -12,7 +12,7 @@ function mul(a: i32, b: i32): i32 {
   return a * b;
 }
 
-function div(a: i32, b: i32): i32 {
+function div(a: f32, b: f32): f32 {
   return a / b;
 }
